Speeds compatibility refactor
